Redirects store
Former-commit-id: 41ff7159c226ae75ea6d1cd4be58a8c6eae6f8b3 [formerly 24a2bfbebbbcae3921a669d256d435d7ac2ae412]
Former-commit-id: 29a9ff1f4f46b08da3be9555ee125018a13dbb2a

+++ api/cache/cache.go
@@ -46,7 +46,7 @@ func ClearUserCache(userID int, posts domain.PostData) {
 	}
 }
 
-func ClearCategoryCache(categoryID int64, posts domain.PostData) {
+func ClearCategoryCache(categoryID int, posts domain.PostData) {
 	for _, v := range posts {
 		if v.Category.Id == categoryID {
 			ClearPostCache(v.Id)
