+++ api/http/handler/api/redirects/find_test.go
@@ -26,7 +26,7 @@ func (t *RedirectsTestSuite) TestRedirects_Find() {
 			http.StatusOK,
 			"Successfully obtained redirect with ID: 123",
 			func(m *mocks.RedirectRepository) {
-				m.On("GetByID", int64(123)).Return(redirect, nil)
+				m.On("GetByID", 123).Return(redirect, nil)
 			},
 			"/redirects/123",
 		},
@@ -35,7 +35,7 @@ func (t *RedirectsTestSuite) TestRedirects_Find() {
 			http.StatusBadRequest,
 			"Pass a valid number to obtain the redirect by ID",
 			func(m *mocks.RedirectRepository) {
-				m.On("GetByID", int64(123)).Return(domain.Redirect{}, fmt.Errorf("error"))
+				m.On("GetByID", 123).Return(domain.Redirect{}, fmt.Errorf("error"))
 			},
 			"/redirects/wrongid",
 		},
@@ -44,7 +44,7 @@ func (t *RedirectsTestSuite) TestRedirects_Find() {
 			http.StatusOK,
 			"no redirects found",
 			func(m *mocks.RedirectRepository) {
-				m.On("GetByID", int64(123)).Return(domain.Redirect{}, &errors.Error{Code: errors.NOTFOUND, Message: "no redirects found"})
+				m.On("GetByID", 123).Return(domain.Redirect{}, &errors.Error{Code: errors.NOTFOUND, Message: "no redirects found"})
 			},
 			"/redirects/123",
 		},
@@ -53,7 +53,7 @@ func (t *RedirectsTestSuite) TestRedirects_Find() {
 			http.StatusInternalServerError,
 			"internal",
 			func(m *mocks.RedirectRepository) {
-				m.On("GetByID", int64(123)).Return(domain.Redirect{}, &errors.Error{Code: errors.INTERNAL, Message: "internal"})
+				m.On("GetByID", 123).Return(domain.Redirect{}, &errors.Error{Code: errors.INTERNAL, Message: "internal"})
 			},
 			"/redirects/123",
 		},
